+++ laravel/resources/views/content/admin/masters/vechiles/types/list.blade.php
@@ -69,10 +69,12 @@
                         <div class="mt-2">
                             {{ $data->onEachSide(5)->appends(request()->query())->links('vendor.pagination.bootstrap-4') }}       
                         </div>
-                        <div class="px-2 my-2">
-                            {{-- deleteMultiple() for delete multiple data pass url here  --}}
-                            <x-admin.form.buttons.multipleDelete url="{{route('masters-vehicle-type-delete-multiple')}}" />
-                        </div>
+                        @can('main-navigation-masters-vehicle-type-delete')
+                            <div class="px-2 my-2">
+                                {{-- deleteMultiple() for delete multiple data pass url here  --}}
+                                <x-admin.form.buttons.multipleDelete url="{{route('masters-vehicle-type-delete-multiple')}}" />
+                            </div>
+                        @endcan
                         <table class="table" id="master-list">
                             <thead>
                                 <tr>
